Conflicts:
	include/clang/AST/Decl.h
	include/clang/AST/OperationKinds.h
	include/clang/AST/Type.h
	include/clang/Basic/DiagnosticSemaKinds.td
	include/clang/Basic/LangOptions.def
	include/clang/Basic/Makefile
	include/clang/Config/config.h.in
	include/clang/Frontend/CodeGenOptions.def
	include/clang/Sema/DeclSpec.h
	include/clang/Sema/Sema.h
	include/clang/Serialization/ASTBitCodes.h
	lib/AST/DeclBase.cpp
	lib/AST/DeclPrinter.cpp
	lib/AST/Expr.cpp
	lib/AST/Stmt.cpp
	lib/AST/StmtPrinter.cpp
	lib/AST/TypePrinter.cpp
	lib/Basic/Makefile
	lib/CodeGen/BackendUtil.cpp
	lib/CodeGen/CodeGenModule.cpp
	lib/CodeGen/CodeGenModule.h
	lib/Driver/Tools.cpp
	lib/Driver/Types.cpp
	lib/Frontend/InitPreprocessor.cpp
	lib/Parse/ParseDecl.cpp
	lib/Parse/Parser.cpp
	lib/Sema/DeclSpec.cpp
	lib/Sema/JumpDiagnostics.cpp
	lib/Sema/Sema.cpp
	lib/Sema/SemaDecl.cpp
	lib/Sema/SemaExpr.cpp
	lib/Sema/SemaType.cpp
	lib/Serialization/ASTWriter.cpp
	lib/StaticAnalyzer/Core/ExprEngine.cpp
	runtime/Makefile
	test/lit.cfg
	tools/CMakeLists.txt
	tools/Makefile
	tools/driver/Makefile

+++ lib/CodeGen/CGValue.h
@@ -468,7 +468,7 @@ class AggValueSlot {
   // Qualifiers
   Qualifiers Quals;
 
-  unsigned short Alignment;
+  unsigned Alignment;
 
   /// DestructedFlag - This is set to true if some external code is
   /// responsible for setting up a destructor for the slot.  Otherwise
